brain/gps/requirements.txt
pyserial>=3.5
pynmea2>=1.18
paho-mqtt>=1.6.1
pyyaml>=6.0
